add the color for navigation buttons
