+++ feature/home/presentation/src/main/java/com/sanaa/presentation/app/NovixApp.kt
@@ -1,11 +1,16 @@
 package com.sanaa.presentation.app
 
+import androidx.activity.ComponentActivity
 import androidx.compose.foundation.background
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+import androidx.core.view.WindowInsetsControllerCompat
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -44,6 +49,20 @@ private fun AppNavigation(
     modifier: Modifier = Modifier,
 ) {
     NovixTheme(isSystemInDarkTheme()) {
+        val navColor = Theme.colors.surface
+        val isSystemInDarkTheme = isSystemInDarkTheme()
+        val view = LocalView.current
+        val activity = view.context as? ComponentActivity
+
+        LaunchedEffect(Unit) {
+            activity?.window?.also { window ->
+                window.navigationBarColor = navColor.toArgb()
+                WindowInsetsControllerCompat(window, view).apply {
+                    isAppearanceLightStatusBars = !isSystemInDarkTheme
+                    isAppearanceLightNavigationBars = !isSystemInDarkTheme
+                }
+            }
+        }
         NavHost(
             modifier = modifier.background(color = Theme.colors.surface),
             navController = AppNavigation.app,
